Moving site-list to its own folder
