+++ client/landing/subscriptions/site-list/site-types.tsx
@@ -1,4 +1,3 @@
-// typo de dato con estos valores: id, name, icon, url, date, emailFrequency
 export type SiteType = {
 	id: number;
 	name: string;
